feature/refactor-easy-encryption
- Refactor EasyEncryption
